Corrected some markdown, spelling, and URL mistakes...
Saw one of @freekmurze 's tweets referring to this (newer) package, which I'd not yet heard of, so I took a look.  I just noticed a few links that weren't working, so came in to make a couple corrections.  I haven't tested the links, but just made what I assume are the right changes. 😜

Please let me know if there are any questions/concerns.

Thanks again for your work...! 🚀